+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+# [fa/agentes] (05/09/2025)
+
+### Changed
+- Separete agent class to other file
+
+# [fa/agentes] (05/09/2025)
+
+### Added
+- POI methods
+- Actions logs list
+- Game win or loss conditions
+
+# [fa/agentes] (05/09/2025)
+
+### Added
+- POI methods
+- Actions logs list
+- Game win or loss conditions
+
 # [Kami/agentes] (04/09/2025)
 
 ### Added
@@ -7,7 +26,11 @@
 - Script for Fire Grid Manager on Unity
 - Route for full State of Grid
 
+<<<<<<< HEAD
 # [fa/unity] (03/09/2025)
+=======
+# [fa/agentes] (03/09/2025)
+>>>>>>> fa/agentes
 
 ### Changed
 - breakWall method to take into account that it must be damages 2 time in order